refactor: move resultTypes.ts file
These types have nothing to do with the parser so they make more sense to be in the src directory (this was the original intent)
