+++ src/resultTypes.ts
